requirements.txt
aniso8601==9.0.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.1
greenlet==1.1.3.post0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.4
python-dotenv==0.21.0
pytz==2022.4
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
